Correct API Token Permission Handling in Backend
The permission handling for tokens was based only on the
username. So a user could only see his own tokens no matter
what kind of permissions he had. Even the admin which should
have full controll over the access could not see or alter the
API Tokens of diffrent users.

* Remove check based on username of request
* Use the RestPermissions check based on the username
* Give a user default permissions for his own API Tokens.