Wire TokenList into UserForm
To list, create and delete the API Tokens we use the new
component and wire the handler with the user store.
We also check if a user is allowed to see and modify
the tokens. For now it is enough to only check if a
user can list the tokens, since if he can do that
he can also modify.